New class DefaultEmailArbitrary.java
